+++ src/core/Attachment.ts
@@ -1,9 +1,17 @@
+import { AttachmentOptions } from "../types.js";
+
 class Attachment {
+  public op: GPUStoreOp = "store";
+
+  public view?: GPUTextureView;
+  public resolveTarget?: GPUTextureView;
+
   constructor(
     public value: GPUColorDict | GPUColor | number,
-    public view?: GPUTextureView,
-    public op: GPUStoreOp = "store"
-  ) {}
+    options: AttachmentOptions
+  ) {
+    Object.assign(this, options);
+  }
 }
 
 export default Attachment;